Move files of game screen to game_screen folder
